// src/containers/DoctorProfile/models/Account.model.ts
import { Role } from "./Role.model";

export type Account = {
    id?: number;
    email: string;
    firstName: string;
    lastName: string;
    ward: string;
    streetAddress: string;
    locality: string;
    city: string;
    postalCode: string;
    phone: string;
    avatar: string;
    dob: string;
    isMale: boolean;
    active: boolean;
    registerTime?: string;
    role?: Role;
};